requirements.txt
ultralytics
Pillow
opencv-python
tkinter
svgwrite
gdown
